requirements.txt
Cython
chnutils>=0.0.7
twisted
pyyaml
git+https://github.com/CommunityHoneyNetwork/hpfeeds3.git
